Added a Veldrid.Desktop solution that converts the most essential projects from netstandard2.0 to .NET 4.6.1.
Please note that parts of the code base use C# 7.1 features such as the modern tuple syntax and thus Visual Studio 2017 update 15.3 or newer is needed to compile the code.

Veldrid depends on some additional .NET standard based mellinoe libraries that are originally pulled in via nuget. Two of these are also converted to .NET 4.6.1 and consumed as git sub modules:
* NativeLibraryloader.Desktop (https://github.com/paralaxsd/nativelibraryloader.desktop) - used to dynamically load libraries in Veldrid.SDL2.Desktop
* System.Runtime.CompilerServices.Unsafe.dll (https://github.com/paralaxsd/System.Runtime.CompilerServices.Unsafe.Desktop) - a CLR4 conversion of https://github.com/dotnet/corefx/tree/master/src/System.Runtime.CompilerServices.Unsafe
=> be sure to clone the repo using --recurse-submodules

Libraries that are currently omitted:
* Veldrid.ImageSharp - There's no quick .NET 4.6.1 replacement for ImageSharp
* Veldrid.ImGui - Depends on ImGui.NET which is also netstandard2.0 only.
* NeoDemo - depends on the first two projects
* Veldrid.Tests - not currently in focus